PM-003 add pizza to menu

+++ src/main/java/com/amazingpizza/api/repository/PizzaRepository.java
@@ -3,5 +3,5 @@ package com.amazingpizza.api.repository;
 import com.amazingpizza.api.model.Pizza;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PizzaRepository extends JpaRepository<Pizza, Integer> {
+public interface PizzaRepository extends JpaRepository<Pizza, Long> {
 }
